+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+export { getCarsByFuelType } from "./getCarsByFuelType";
+export { getCOEResultByMonth } from "./getCOEResultByMonth";
+export { update } from "./update";
